+++ myems-web/src/components/MyEMS/Space/SpaceEnergyCategory.js
@@ -38,18 +38,16 @@ import AppContext from '../../../context/Context';
 import { useLocation, Link } from 'react-router-dom';
 import blankPage from '../../../assets/img/generic/blank-page.png';
 
-
-
 const ChildSpacesTable = loadable(() => import('../common/ChildSpacesTable'));
 const DetailedDataTable = loadable(() => import('../common/DetailedDataTable'));
 const WorkingDaysConsumptionTable = loadable(() => import('../common/WorkingDaysConsumptionTable'));
 
-
 const SpaceEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
   let current_moment = moment();
   const location = useLocation();
-  const uuid_ = location.search.split('=')[1];
-  const { isDark } = useContext(AppContext);
+  console.log('location=' + location.search.split('='));
+  //parse space uuid from url, the space uuid is null in most cases, and it exists only when clicking on the map
+  const spaceUUID = location.search.split('=')[1];
   useEffect(() => {
     let is_logged_in = getCookieValue('is_logged_in');
     let user_name = getCookieValue('user_name');
@@ -208,10 +206,14 @@ const SpaceEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
               .split('"name":')
               .join('"label":')
           );
+          // initilize space cascader
           setCascaderOptions(json);
+          // select root space name
           setSelectedSpaceName([json[0]].map(o => o.label));
+          // select root space id
           setSelectedSpaceID([json[0]].map(o => o.value));
-          loadUUID(json);
+          // load data with space tree
+          loadWithSpaceTree(json);
         } else {
           toast.error(t(json.description));
         }
@@ -221,10 +223,11 @@ const SpaceEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
       });
   }, [t]);
 
-  const spaceName = (spaceList, spaceID, name, ids) => {
+  // recursive parse space full name from space tree
+  const spaceFullName = (spaceTree, spaceID, name, ids) => {
     let newName = name;
     let newIds = ids + '';
-    for (let space of spaceList) {
+    for (let space of spaceTree) {
       if (space.value === spaceID) {
         setSelectedSpaceName(newName.length === 0 ? space.label : newName + '/' + space.label);
         let idArr = [];
@@ -235,7 +238,7 @@ const SpaceEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
       }
       let rt = false;
       if (space['children'] && space.children.length > 0) {
-        rt = spaceName(
+        rt = spaceFullName(
           space.children,
           spaceID,
           newName.length === 0 ? space.label : newName + '/' + space.label,
@@ -246,13 +249,15 @@ const SpaceEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
     }
     return false;
   };
-  const loadUUID = spaceList => {
-    if (uuid_ !== null && uuid_) {
+  const loadWithSpaceTree = spaceTree => {
+    // there is spaceuuid in the url
+    if (spaceUUID !== null && spaceUUID) {
+      // make api url with spaceuuid
       let url =
         APIBaseURL +
         '/reports/spaceenergycategory?' +
         'spaceuuid=' +
-        uuid_ +
+        spaceUUID +
         '&periodtype=' +
         periodType +
         '&baseperiodstartdatetime=' +
@@ -265,11 +270,36 @@ const SpaceEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
         moment(reportingPeriodDateRange[1]).format('YYYY-MM-DDTHH:mm:ss') +
         '&language=' +
         language;
-      loadData(url, spaceList);
+      loadData(url, spaceTree);
+    } else {
+      // there isn't spaceuuid in the url
+      // make api url with spaceid
+      if (spaceTree && spaceTree.length > 0) {
+        let spaceID = [spaceTree[0]].map(o => o.value)[0];
+        let url =
+          APIBaseURL +
+          '/reports/spaceenergycategory?' +
+          'spaceid=' +
+          spaceID +
+          '&periodtype=' +
+          periodType +
+          '&baseperiodstartdatetime=' +
+          (basePeriodDateRange[0] != null ? moment(basePeriodDateRange[0]).format('YYYY-MM-DDTHH:mm:ss') : '') +
+          '&baseperiodenddatetime=' +
+          (basePeriodDateRange[1] != null ? moment(basePeriodDateRange[1]).format('YYYY-MM-DDTHH:mm:ss') : '') +
+          '&reportingperiodstartdatetime=' +
+          moment(reportingPeriodDateRange[0]).format('YYYY-MM-DDTHH:mm:ss') +
+          '&reportingperiodenddatetime=' +
+          moment(reportingPeriodDateRange[1]).format('YYYY-MM-DDTHH:mm:ss') +
+          '&language=' +
+          language;
+        loadData(url, null);
+      }
+
     }
   };
 
-  const loadData = (url, spaceList) => {
+  const loadData = (url, spaceTree) => {
     // disable submit button
     setSubmitButtonDisabled(true);
     // show spinner
@@ -300,10 +330,9 @@ const SpaceEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
       })
       .then(json => {
         if (isResponseOK) {
-
-          if (spaceList && spaceList.length > 0) {
+          if (spaceTree && spaceTree.length > 0) {
             let spaceID = json['space']['id'];
-            spaceName(spaceList, spaceID, '', '');
+            spaceFullName(spaceTree, spaceID, '', '');
             setSelectedSpaceID(spaceID);
           }
           let cardSummaryArray = [];